Update Thrift's Hack codegen to use optional and unknown shape fields when generating code
Summary:
= This change

Thrift is a codegen framework that exposes auto-generated code which uses shapes. I've updated the tool to incorporate the new shapes syntaxes (see Background, below).

This change doesn't do anything smart, it simply preserves the existing semantics in terms of the new shape syntaxes. If you'd like to do something smart (e.g., fields may be absent but will not be null, or something like that), let me know; or, you can do it yourself in a future codegen modification if you think it will be a complex or error-prone change.

= Background

Reviewed By: eduardo-elizondo

Differential Revision: D5436934

fbshipit-source-id: 42a1e41df3461faa6261d96df29a5f28f23eacdc