Allow const access to bundled holders
Summary:
This adds a very restricted interface to access objects held by `Bundled`
instances. The limitations are:

- It's only ever possible to get `const` access to held objects
- If there are multiple held objects of the same type, you can only get
  access to the first object

This should still be sufficient for a lot of use cases and is much easier to
use (and less dangerous) than allowing direct access to the holder vector.

The motivation for adding this is that it allows to get (read) access to the
`folly::MemoryMapping` instance that's held by `MappedFrozen` if it was
mapped from a file. This in turn allows, for example, to get (in most cases)
to the name of the file that was mapped, which is extremely valuable for
logging errors that arise when all we have access to is the `MappedFrozen`
instance.

Reviewed By: yfeldblum

Differential Revision: D5527160

fbshipit-source-id: 3d34dd06023c1f6141e51558d8dabacde8d74ac3