Introduce PeekingManager
Summary:
This renames the SSLRejectionManager into PeekingManager as it will do more than just peeking for SSL connections.

For now, this will error out HTTP2 connections if they are sent over plaintext messages.

Note, this codepath is only executed if we have plaintext messages. If we are using TLS this will go through a completely different paths in which we will use APN to determine the client type.

Reviewed By: phoad

Differential Revision: D5616009

fbshipit-source-id: 4cdd13f7c17834a6ce8b3e45dd623202f2655385