Make TSSLSocket for Python clients support latest TLS
Summary:
TSSLSocket will detect whether or not the Python environment can be configured
with the newer Python 2.7.9+/3.2+ APIs. This enables TSSLSockets to speak
newer versions of TLS.

By default, if the user does not specify a SSL version to use, we will configure
TSSLSockets with TLS 1.1+, matching PCI compliance.

If an old, legacy Python platform is detected, a warning message will be logged.

Reviewed By: ngoyal

Differential Revision: D5599248

fbshipit-source-id: 25ed63b0512c0a0adc43c8415ce16959f6657ba1