Compiler: Create new objects for constant enums
Summary:
There's cases where we have:
```
enum AnEnum { FOO = 1 }
const i32 const1 = AnEnum.FOO
struct AStruct {
  1: AnEnum const2 = AnEnum.FOO
}
```

These two constants share a pointer to the const value of foo. If we want to support the case where we have integers receiving enums, we need to modify the t_const_value to point to an int type.

This creates a copy of the const value so that it enables us to do this.

Reviewed By: yfeldblum

Differential Revision: D5647183

fbshipit-source-id: 997d59d34fd39d8e6ced58cf89b39d70f2f2ee7d