Handle cpp.template type annotation
Summary:
Add support for the cpp.template annotation on types. When this annotation is provided, we generate the appropriate extern call with the minimum set of required methods to include the implementation class, and then everywhere we would reference (eg) std::map, we reference the other class instead. If the implementation class is in the library already (like std::map), we change the reference but don't do the extern statement. This implementation subsumes the earlier work to support unordered sets and maps.

It is possible there are still some rough edges here but the basics work and I thought it better to just get it out and let people start using it.

Reviewed By: yfeldblum

Differential Revision: D5624254

fbshipit-source-id: a0dd34d60c443ac44ddbd0a73684dd0899c59c9c