Fix makefile breakage
Summary:
The addition of TransportRoutingHandler.h in D5655353 is breaking open-source builds. The reason is this:

```
/usr/bin/install -c -m 644 server/BaseThriftServer.h server/Cpp2ConnContext.h server/Cpp2Connection.h server/Cpp2Worker.h server/ThriftServer.h server/peeking/PeekingManager.h server/peeking/TLSHelper.h server/peeking/HTTPHelper.h transport/core/TransportRoutingHandler.h '/home/install/include/thrift/lib/cpp2/server'
```

Note the target is: `cpp2/server/transport`.

But we are looking for it in `cpp2/transport`.

```
/home/install/include/thrift/lib/cpp2/server/ThriftServer.h:46:68: fatal error: thrift/lib/cpp2/transport/core/TransportRoutingHandler.h: No such file or directory
```

I believe that it was a mistake to add it to `thrift2include_server_HEADERS`, so I am trying to monkey the pattern to put it into the appropriate directory here.

Reviewed By: yfeldblum

Differential Revision: D5677443

fbshipit-source-id: f56896848463f7f067109427698eef5d65b7bdbc