Use vector of string for arguments
Summary:
This modifies main.cc to use a vector of strings rather than argv. This will allow us to be able to modify the generator strings when needed. If we need to call the python generator we convert back to char** based on the modified generator strings.

This is a pre-step to be able to generate both mstch_cpp2 structs and cpp2 reflection structs.

Reviewed By: yfeldblum

Differential Revision: D5677712

fbshipit-source-id: 7bbc395e272620b42109476752aa2a2c8de47343