mstch-cpp2: Deep copy const_values
Summary: Add a deep copy constructor for const_values. This expands on the work of using a mutator. There are cases where a const value gets assigned to a type that differs in integers enums. By using this deep copy, these two const values are independent of each other and can be mutated accordingly.

Reviewed By: yfeldblum

Differential Revision: D5729178

fbshipit-source-id: 8dc2d4d0ef6c4c7b81e0e67da82581f476b2b480