Uncomment setInput / getOutput function in the main interface
Summary:
As we already depend on an alias, which is SubscriberRef, if we change the definition of this alias, we will not have any compilation failure as we will not have any dependency.
We can in the near future make it refer to yarpl::Subscriber. So I have replaced its definition and uncommented rest of the already unused code.

 ---

As Thrift now depends on RSocket also, all the projects that depend on Thrift also depends on RSocket.
Some of these projects use fbconfig/fbbuild as the build tool, which misses some functions that buck has.

Two of them are 'subdir_glob' and 'header_namespace'. It is easy to mimic subdir_glob but it was not possible to hack the 'header_namespace'.

We will have another update in RSocket which will update the folder structure, but for now we can uncomment some of the functionality to be updated later.

Reviewed By: eduardo-elizondo

Differential Revision: D5715173

fbshipit-source-id: da9c23b71680900ebf200cf245d3e046a0290fa6