Add support for having separate IO and Worker threads
Summary:
This is the change that enables having IO and Worker threads.
Before this change even having single IO Thread was causing crash, checkEventBase was crashing.
If we give EventBase, because it is not attached to any thread, checkEventBase function simply returns true, it was returning false if the eventbase was coming from an IOThread.

Reviewed By: eduardo-elizondo, sriramsankar

Differential Revision: D5715182

fbshipit-source-id: 792c24561405aa7c9da8cd29d474da48f51f8c26