Implement stream/client throws in cpp2 *.mustache
Summary:
A proof of concept for what the client/server interface for pubsub in Thrift
could look like. Not meant to be merged, but to demonstrate areas of the
compiler that need to be modified.

Findings: Thrift's design is pretty complex, going to likely need
some pretty large changes to how requests/responses are made and how
state is kept around on the client. We'll need to consider how
the processors handle async stream elements coming from iothreads.
There's currently no API to support this.

A the moment, generated code compiles, but does not link

Reviewed By: eduardo-elizondo

Differential Revision: D5451079

fbshipit-source-id: ed02d1b7ef90987b4dcbb453f17f1ab6cc3397c7