add fixture to highlight issue with transitive imports in go thrift
Summary: When package A imports B, which imports C, and A defines A.f, which includes B.g and has a default set to B.defaultG, and B.f includes C.h, with a default set to C.defaultH, the thrift compiler would try and build a literal in package A that created a C.h literal. This would fail, as A does not import C directly. This diff adds a fixture exhibiting this behaviour

Reviewed By: nemith

Differential Revision: D5754861

fbshipit-source-id: 627d13dee15c8081594bbda7f6b517e7bf8142f2