fix issues with doubly nested imports
Summary: When package A imports B, which imports C, and A defines A.f, which includes B.g and has a default set to B.defaultG, and B.f includes C.h, with a default set to C.defaultH, the thrift compiler would try and build a literal in package A that created a C.h literal. This would fail, as A does not import C directly. This diff changes the logic to have A simply call B.NewG, which can then call C.NewH.

Reviewed By: yfeldblum

Differential Revision: D5754917

fbshipit-source-id: 3e05264852e95a865e06a11b0f7c4f6f8de187b6