Second diff to support client side code
Summary:
The first diff firmed up the overall architecture and allowed Fuat to
get the RSocket example fully running.  This diff fills the remaining
holes and also addresses some of the error cases.  There will be
smaller cleanup diffs after this one, but this completes the bulk of
the heavy lifting on the client side.

Reviewed By: phoad

Differential Revision: D5761409

fbshipit-source-id: 104217658430955243a9e85ddb5e89026c730835