mstch-cpp2: Refactor forward declares, typedefs, and enums in own scope
Summary:
Previously we defined some rules of when to add a forward declare and when to not.
This modifies that to add forward declares and typedefs in an ordered way an in their own namespaces to avoid complicated logic.

Also, this modifies enums to only use one scope instead of one for every enum.

(Note: this ignores all push blocking failures!)

Reviewed By: yfeldblum, sriramsankar

Differential Revision: D5754259

fbshipit-source-id: 910750bab9685c9cc35253b49b62b8d34220e9ef