Add EvbHandshakeHelper
Summary:
EvbHandshakeHelper is a wrapper around an existing handshake helper that
executes handshake related methods (start(), dropConnection()) in a separate
EventBase thread that is specified upon creation time. Callbacks are delivered
in the original thread that the caller start()ed the helper in.

This can be used with existing TLS accepting helpers to offload expensive
handshakes onto a separate thread, which increases handshake throughput.

Reviewed By: siyengar, knekritz

Differential Revision: D5682681

fbshipit-source-id: 6ef885b774d3b83b3f2156ed0906bb175eabc0f6