Handle exception during thrift serialization
Summary:
Currently when we fail during thrift serialization, the application is not
notified of the exception. Although these exceptions are available in the
stderr output, these don't get logged. Due to this it takes longer to debug
issues when they happen.

Also client side continues to wait for response till it times out. Now we close
the transport causing the client to immediately detect problem. Not that the
client will only get a "Connection Closed" error and not the actual exception
that caused the connection to close.

Reviewed By: yfeldblum

Differential Revision: D5768672

fbshipit-source-id: 70a4a72b8636d2817b9409de6cfb1198cfc3afa7