Delaying calling getTransaction to immediately prior to sendHeaders
Summary:
In Proxygen, getTransaction assigns a stream id, but the stream is actually created in the transport only after sendHeaders is called.  This sometimes caused the stream ids to be out of order in the transport.

This diff moves the call to getTransaction right before the call to sendHeaders.

In addition, added some more DCHECK's and also changed SingleRpcChannel to use the connections event base (which could change over time).

Reviewed By: phoad

Differential Revision: D5819628

fbshipit-source-id: 7622680516e8ea64a50522ecea77884b899cebb8