Compression: Properly ifdef Snappy lib
Summary:
thrift/lib/cpp/thrift_config.h defines: HAVE_LIBSNAPPY. However, this is not used to properly guard the snappy code. Plus, we added the definition in the CMake file to remove it from the open source build as we want to minimize external deps.

Ideally, we should create a thrift_config.h.in file for cmake to automagically generate the thrift_config.h.

Reviewed By: yfeldblum

Differential Revision: D5826060

fbshipit-source-id: 98e19ff051cfa292fb9c148aa1b538ecb0e69e9d