oss: Enable thrift open source example
Summary:
This enables an executable that will require the linking of the thrift libraries. The build passing demonstrates that the linking successfully passed.

Note: there's a known bug in LLVM/Clang in which "___cxa_deleted_virtual" is not defined. See: https://reviews.llvm.org/D36870. Thus, this only works with the gcc compiler.

Note2: In mac you also need to use the openssl  v1.0.2g

Note3: This temporarily disables Zlib in open source. The reason is that the current build system in sandcastle has a really old cmake that can't properly find Zlib. I'm migrating this to a docker build which should have the required libraries.

Reviewed By: yfeldblum

Differential Revision: D5826080

fbshipit-source-id: cd8ad5b68a38810424e4408e83048b9fb00d369d