Fix the TLS related bug in ConnectionThread.cpp
Summary:
Even though we create either H2 client, or RSocket client, in the TLS negotition, sslContext, we say that we are advertising all kinds of protocols. So the server selects H2 over RSocket.

So I have updated the code to say that if we have RSocket client, then only promote `rs`, otherwise promote `h2` or `http`.

Reviewed By: eduardo-elizondo

Differential Revision: D5887413

fbshipit-source-id: 9616395d93642264e4c163e5ad3aff1af294e079