In Memory transport for tests/debugging
Summary:
This diff implements a transport that connects a Thrift client and
server together within the same binary.  It uses the same API as the
HTTP2 and RSocket transports.  Both clients and servers can be of any
kind (sync, async, or future).

This can be used as a unit testing framework to combine Thrift clients
and servers into a single binary.

Reviewed By: phoad

Differential Revision: D5928798

fbshipit-source-id: edec20bc87c195ee6b7eb758f333cb57f00570b9