Move some generated client code to GeneratedCodeHelper
Summary: There is some code to throw an ApplicationException if the buffer holds an unknown ApplicationException, if the method name of the response is incorrect, or if the response is neither an exception or reply. This code does not need to be generated, and we can reduce on code size by factoring it out.

Reviewed By: yfeldblum

Differential Revision: D5980006

fbshipit-source-id: 54474e1b3269ae49fe6308fe725bb523f5f0b137