Extend the Service to have streaming enabled function.
Summary:
Added some extensions to the compiler generated code.
I made it possible to make streaming enabled (Channel) Thrift RPC calls.
These changes will be integrated to the compiler.
It is costly to do the changes in the compiler, so I have hand written (copy/paste) the code to figure out the expected compiler updates and test the transport layer.
As the compiler updates get ready, these generated code should be removed and should be replaced by auto generated code.
So we will be able to easily test the compiler generated code.

Differential Revision: D5952687

fbshipit-source-id: 0c90928456faa6bcdf45f4fff24998b46f3f464b