Make thrift enums use IntEnum
Summary:
Enums in thrift are integers. Using python's IntEnum
(https://docs.python.org/3/library/enum.html#intenum) as
the baseclass allows to use enums similarly to previous python
versions.

This code was working in previous py2 and py3:

  SomeEnum == 3
  SomeEnum.FOO > SomeEnum.BAR

If we use regular Enum, you have to use `.value` to get the integer
version if you want to compare with an int or use `>` & `<` operators.

Reviewed By: danshiovitz

Differential Revision: D6000793

fbshipit-source-id: 3c60e8dca453f245f2d88160771ade3cd9acfefd