Rename context object to avoid name clashes
Summary: There's a possible name collision here if the thrift service has a method that has a parameter named 'context' or 'promise'. This change renames them to something less likely to collide, but we'll probably need a more aggressive name-mangling plan for the future, and possibly should be mangling the names of the user's arguments instead of our own parameters.

Reviewed By: yfeldblum

Differential Revision: D6013789

fbshipit-source-id: 9d3b1b95103773a5205b3006288f011626ec539f