Move non-presult Cpp2Ops to its own file
Summary: Cpp2Ops is not just used in generated code - it's also used in DebugProtocol and Serializer. It just so happens that currently, everyone is including GeneratedCodeHelper headers with their structs, but I plan to change that in a subsequent diff. Therefore, Cpp2Ops needs to be its own header and tcc, and include those in the places that need them today.

Reviewed By: eduardo-elizondo, yfeldblum

Differential Revision: D6052849

fbshipit-source-id: 4fc1180f7b4f7788403f5c433fd74d50630240fc