Remove envelope from ThriftClient when channel supports
Summary:
Removes sending the envelope for channels which understand
how to use RequestRpcMetadata, instead of duplicating this
information in the envelope

We're effectively deciding that the sequence ID is hardcoded to 0 in `clientSendT`, which doesn't seem correct. This doesn't change that, but it does add another place where that's hardcoded in `RSClientThriftChannel` (which would normally get that information from stripEnvelope).

Reviewed By: phoad

Differential Revision: D6056018

fbshipit-source-id: ae802c65a73308b4e8d789927ad4b60aab7f71d4