Introduce ServerConfigs interface to prevent cyclic reference between BaseThriftServer and ThriftProcessor
Summary:
Previously, ThriftServer/ThriftBaseServer and Cpp2Connection classes were part of the same project. So even if they have had cyclic reference, because they were in the same project, it was ok to have this cycle.

Now, as we introduce new transport layers, ThriftBaseServer and ThriftProcessor (replacement for Cpp2Connection) classes exists in two different projects. The project that contains ThriftBaseServer already references the project that contains ThriftProcessor.

"Server Timeout" and "Max Request Size" related functionalities requires ThriftProcessor class to ask the current limits from the ThriftBaseServer object. So it requires access to ThriftServer also. This has required the project which contains ThriftProcessor to have a reference back to the project that contains ThriftBaseServer. Which causes build failure of having cycle between projects.

To break this cycle, I have introduced ServerConfigs interface that ThriftBaseServer extends to. I have also separated the project that holds ServerConfigs from the others. So both of the projects which keeps ThriftBaseServer and ThriftProcessor can freely reference this new project and cyclic dependency problem gets resolved.

Reviewed By: sriramsankar

Differential Revision: D6133035

fbshipit-source-id: ab0f91b581cdd3b7887481dd5a569f170aed04a3