Perf: Cleanup Operations
Summary:
Modified the Perf benchmarks for easier extensibility. The structure is as follows:

Client.cpp creates a Runner object per thread.
The Runner object owns a client, an outstanding operations counter, and a random number generator to choose any given operation depending on it's discrete distribution based on the weights set by the Flags.

If a new operation is added to Api.thrift, then it's just a matter of creating an object in SimpleOps.h, extending the enum in Operation.h, and adding the weight in Client.cpp.

Reviewed By: phoad

Differential Revision: D6127546

fbshipit-source-id: 99f2e864d5a41fb40859eb67158df1fbbba6d9e3