Introducing Client Proxies
Summary:
To allow for more robust testing mechanisms we need to have a way for returning
proxies from get_client so certain methods could be mocked out.

To insure get_client's are not called before we get a chance to install a proxy
factory I have added a check that the eventloop is running.  This will
prevent the use of get_client at module level (at import) and used later, in
which case the proxy would not have been used.

This is not a major change in that it was always expected that get_client was
called from an 'async def'

Differential Revision: D6116067

fbshipit-source-id: 35418f872d381b7cb266ced7caf8dcdec578ed69