MetadataInBodySingleRpcChannel
Summary:
This commit introduces MetadataInBodySingleRpcChannel which inherits from SingleRpcChannel but differs in the way RequestRpcMetadata is serialized over the wire. MetadataInBodySingleRpcChannel will serialize RequestRpcMetadata as thrift compact structure and send in body instead of converting to strings and sending in header.

For now, the behavior is toggled by the thrift_cpp2_metadata_in_body commandline flag.

Reviewed By: eduardo-elizondo

Differential Revision: D6138491

fbshipit-source-id: 9f75b0663f0a1f38f7978403af7f190af084eece