Cleanup and fixes for Service Router integration
Summary:
The update in the Service Router's ConnectionManager seemed buggy about the SSL check (D6113118).
This diff should be fixing that.

getSecurityProtocol() is defined as
```
  std::string getSecurityProtocol() const override {
    if (sslState_ == STATE_UNENCRYPTED) {
      return "";
    }
    return "TLS";
  }
```
in the folly::AsyncSSLSocket and it overrides this functionality in the folly::AsyncTransportWrapper which returns empty string.

So we should depend on this function's value to decide if the transport is `secure` or not.

We cannot assume the transport is secure.

Reviewed By: GeniusYe

Differential Revision: D6179280

fbshipit-source-id: 3c9605e5d62f0e8f06c1c31f233df01a62eed144