Fix http handling of oneway
Summary: HTTP/1.x servers (e.g. HHVM) will respond to oneway requests with an empty 200 response. This is breaking HeaderClientChannel, because it is passing the empty result to the next two-way callback. In addition, THeader is clearing the entire IOBufQueue when a single request is parsed, which is incorrect. It should only trimStart() the request that has been parsed.

Reviewed By: eduardo-elizondo

Differential Revision: D6182480

fbshipit-source-id: 324fa5167af8b2c584038aaa86d7076cf41ed8f5