Handle the case if the connection gets closed internally
Summary:
If the client is an RSocket client but the server doesn't talk in RSocket frames, then the server drops the connection.

Without this update, even though the connection is dropped, the code was assuming that the connection is still available.
Now, it first checks if the connection still exists.

Differential Revision: D6240014

fbshipit-source-id: beefc084b91bfc4e84f6756458fcb654a10e4612