ProtocolReader::kUsesFieldNames
Summary:
[Thrift] `ProtocolReader::kUsesFieldNames`.

So that cpp2-generated deserialization code can tell whether it is required to translate field names by looking at the protocol reader, rather than at the field-id.

And since the deserialization code is templated over the protocol readers, if the particular protocol reader implements the function as `constexpr`, the instantiated deserialization code can optimize away the check and the call.

Reviewed By: eduardo-elizondo

Differential Revision: D6244740

fbshipit-source-id: fe6798a2df66b4af951979eb72e25e32b69aab61