Transport: Allow the Client owner to set an HTTP Host and URL
Summary: Previously, http host and url was set only by the transport. There are use cases in which we don't know the correct host and url at creation time. Thus, we need a way to be able to dynamically set the URL and Host. This enables ThriftClient to be able to get the connection to dynamically set the correct Host and URL values.

Reviewed By: sriramsankar

Differential Revision: D6311857

fbshipit-source-id: c8d4db505df22c7dbe966e68413b3e1bc2f396f6