Outline destructors of large structs
Summary:
Follow-up to D6242284 for destructors, which are actually larger than constructors (except move constructors) because they may have to deal with deallocation.

To ensure there is no performance penalty we tweak the "large" criterion: beside the threshold on number of members, we require that the struct has at least one "heap-allocated" member (strings, containers, refs), so we are sure that the destructor is non-trivial and has at least one branch (and likely a call to `free`).

Reviewed By: yfeldblum

Differential Revision: D6310143

fbshipit-source-id: e57cbc245a2470aaa6e38230d2f801e8c445ac13