Unbreak thrift OSS build
Summary:
Open source thrift was missing a critical library and binaries that linked
against it would get:

../external/fbthrift/build/lib/libthriftcpp2.a(ThriftProcessor.cpp.o): In
function
`std::default_delete<apache::thrift::RequestRpcMetadata>::operator()(apache::thrift::RequestRpcMetadata*)
const':
ThriftProcessor.cpp:(.text._ZNKSt14default_deleteIN6apache6thrift18RequestRpcMetadataEEclEPS2_[_ZNKSt14default_deleteIN6apache6thrift18RequestRpcMetadataEEclEPS2_...

The immediate problem was introduce in D5911253 but:

The core problem related to misunderstanding about the scope
of Cmake variables.  When you define a variable in directory,
that variable is accessible in any subdirectory, but not in the parent
or sibling directory.  This patch fixes the immediate problem,
but there is a more structural problem that needs to be addressed.

Reviewed By: spalamarchuk

Differential Revision: D6392117

fbshipit-source-id: e03aa568966cd37ed35867d8a76d3633656830b6