Don't generate empty service code
Summary:
We had an edge case where old thrift files defined empty services.
Because we generated no api service methods cython did not generate a _api.h
file.

I can see a possible use-case for empty inheritence of an existing service but with
custom annotations.  And there are empty service definitions in existing thrift files.

So in the case where no callback service methods were defined we do not generate
the include of the _api.h (We can get rid of the _api.h all together in the future when
we replace our ServiceHandlers with a folly::function based registry)

Also there is no reason to generate empty service and client code.

Depends on D6381275

Reviewed By: yfeldblum

Differential Revision: D6387668

fbshipit-source-id: 5f2b8caba7faabd20a87e80aac2f55a644d37d7e