Mangle All imports so shadowing is harder
Summary:
We had edge cases where identifiers defined in thrift files could
conflict with import names and screw up cython compiling.

To fix this all imports are now imported as _thing with all . replaced with _

This required a few changes to the compiler, I made get_py3_namespace accept a
list of tails to append to the array we pass to mustache, before we had ".types"
hard-coded to append to the program_name, this prevented the template from
generating correct _ ModulePath.

Reviewed By: yfeldblum

Differential Revision: D6381275

fbshipit-source-id: eb80043b17b788c102f00d7bfff68ba04d512ac1