Acceptor: Fix Plaintext To TLS Server Acceptor
Summary:
There was an incorrect assumption that was added to Thrift. This was that all non-tls clients would go to plaintextConnection. However, this is not true. If the server is a TLS server, it should still accept non-TLS clients.

Before, plaintextConnection was performing the checks to reject TLS connections or accept and route HTTP2 and RSocket connections (plus any other that could be added in the future).

Now, this changes the logic to bring plaintextConnection back to only reject TLS connections and in "onNewConnection" is where we do the routing logic. This now checks the bytes to determine if it's rsocket or http2, if not, then it will just fall through and perform a header operation.

This will keep the same logic for TLS Client -> TLS Server, TLS Client -> Server, and Client -> Server. The benefit is that this now enables Client -> TLS Server.

Reviewed By: phoad

Differential Revision: D6399236

fbshipit-source-id: 209b0c7dff1186942ab65fe49f4da2b63049a2b5