Replace yarpl::Flowable with apache::thrift::Stream
Summary:
Using `yarpl::Reference<yarpl::flowable::Flowable<T>>` in places may hinder our chance to replace it easily in the future and there is no need to be such specific.
So I have replaced it with the generic `apache::thrift::Stream` type.

So the declaration will be generic and prettier.
`apache::thrift::Stream<T> myStreamFunction(..`

Reviewed By: eduardo-elizondo

Differential Revision: D6433939

fbshipit-source-id: 19586385056b9f6f8e8c5ed8e37a8218430d2172