thrift: cpp2: don't sign extend MessageType in compact
Summary:
In compact protocol, the message type is a 3 bit field. Treat it as
unsigned so that shifting after masking doesn't result in sign extension.
Without this, OneWay messages are decoded as -4 instead of 4.

It is implementation-defined whether right-shift of a signed value is
sign-extending, so this may work as expected sometimes.

It turns out in practice this change is moot, because OneWay is never used.
However this change will be necessary if messagetypes >= 4 are ever used.

Reviewed By: yfeldblum, Imxset21

Differential Revision: D6337572

fbshipit-source-id: baaf408af8bd0b09655bc2dc80f0e804186410dc