Save processor factory for InMemoryConnections
Summary:
`newInMemoryClient` takes a `shared_ptr<ServerHandler>`, which is used to build a processor factory, which in turn builds a processor using the raw pointer inside the `shared_ptr`.  This doesn't increment the reference counter for the pointer, so if the original `ServerHandler` goes out of scope it can be destroyed before the client is used.

Since the processor factory has the pointer to the `ServerHandler`, this patch just saves a copy of that factory.  This keeps an open reference to the shared pointer, and prevents it from getting destroyed if the caller's pointer is discarded.

Reviewed By: yfeldblum

Differential Revision: D6430101

fbshipit-source-id: ebfa14e2bf405a9bc0baf0a75a23d3c4632a8a93