Add support for task priorities
Summary:
Adds support for task priorities in TCppServer, which is used by SeRF, which has issues with not being able to answer getStatus calls in a timely fashion when the requests start accumulating in the task queues.

The current implementation ignores method priorities, so everything piles up in the `NORMAL` queue.

Now annotations in the thrift interface like `(priority = 'IMPORTANT')` will be taken into account when scheduling requests, as well as the `thrift_priority` header.

Reviewed By: yfeldblum

Differential Revision: D6428089

fbshipit-source-id: 342fd2f2c4c7c9e05454293f56648760637df403