Add Server Tests and fix pass_context_<rpc>
Summary:
so the type aware pass_context_ decorators lie about the
return type to make mypy happy that works as long as you don't
also call those handler methods from your tests.  So instead
the decorators return a type union of the two call styles.

This is a lie in that the non-context pattern can not actually be called.
But the context pattern still works.

I discovered a bug in services.pyi where a ServiceInterface extends a Service in the same file. We fixed this bug in the .pxd and the .pyx and now the .pyi
This is a limitation of mypy that we have to work around.

Reviewed By: yfeldblum

Differential Revision: D6490078

fbshipit-source-id: 88d2af589fbfdc1a9257c465d6d5254e4c0f3f7d