Http2: Fix oneway requests
Summary:
Currently, every request that was sent was set to SINGLE_REQUEST_SINGLE_RESPONSE. This actually segfaulted the tests when running under ubsan.

This adds the ability for the thrift client using legacy mode, that is, still using THeader to pass in rpckind and to read that on the server side to correctly handle oneways.

Furthermore, this removes the legacy compatibility one way requests. The original "jiejie's" client, is currently not sending this "rpckind" information in the header, this the server will never be able to detect if it's a oneway or a request response.

Reviewed By: sriramsankar

Differential Revision: D6495316

fbshipit-source-id: 9062b7a35b5e778ddac35e605dfbfe262fd7d2b8