Better multi-threading support for Stream RPC calls
Summary:
It is possible to use the same Connection object for multiple ThriftClient instances.
As an optimization when we call getChannel() from the Connection, we return the same Channel object, as the Single Request Single/No Response RPC calls don't depend on anything at the ThriftClientChannel objects.

So we also need to make the Stream RPC calls be independent of each other. To achieve this and to cleanup the ThriftChannelIf interface, don't place anything related to a single Stream request to the channel but keep it at the StreamRequestCallback object always.

Reviewed By: eduardo-elizondo

Differential Revision: D6493355

fbshipit-source-id: ebaabf28298883e4db4e6f835f9930e96fdcbf8a