Cleanup types for pass_context_ again
Summary:
pass_context_ decorators make defining methods that take context not raise type
erros, but they don't protect callers from unittests that pass in context.

This diff reverse the change to pass_context to try and make the output of those
functions union of two callable patterns, because it didn't work.

For now the best bet is to just #type: ignore calls to handler methods with context in
unittests until we find a better solution.

Reviewed By: yfeldblum

Differential Revision: D6501192

fbshipit-source-id: fe0c437f11af0f13c72daa4c6735356bc94161ee