Improve unpickling support for structs
Summary:
If you pickle an object using an old definition of a Thrift struct and
unpickle it with a new version that has added fields, then those fields won't
be defined on your unpickled object. We can fix this by implementing
`__setstate__` and having it set the default values for missing attributes.

Reviewed By: yfeldblum

Differential Revision: D6464541

fbshipit-source-id: 781d0459e169d1af490a8ed28c608a04a89da680