Unions Enhancements
Summary:
Remove unneeded property's from Unions for type and value and instead
just use "readonly" This simplifies the code quite a bit and should be a little
faster

Also provide a fromValue helper that can take a value and attempt to construct
a union by picking the right field automaticly
We try to make the smart decision if multiple integer types are in the union
and multiple floating point types.

If the union is setup from smallest to largest we will always pick the propper
size.

For unions that contain doubles and float we will pick double over float if
there is any loss in precision.

Finally if there are no int fields we will attempt to shove them into float or
double

Reviewed By: danshiovitz

Differential Revision: D6553819

fbshipit-source-id: 56a89063d2ed07350121bf3ed39d4522ba169dba