pyi typechecking generation fixes
Summary:
This generator and templates was copied from thrift-py3 and thus
had some of the same flaws.  I fixed up the compiler so we could pass
the t_program to the service.pyi,  and added the service in question into
a special {{#service}} block.

Also fixed blocks that referenced 'py3'

Reviewed By: yfeldblum

Differential Revision: D6580847

fbshipit-source-id: 4b2cf9b31b75434852585de5cc4ab6b8b4c45c1d