Enums as flags annotation
Summary:
Some use cases of enum thrift types is as a bitmask.
This adds support for that use case when the annotation "py3.flags" is
added to Enum definitions

This makes use of the Enum.Flag class which is pretty robust and allows
and value that can be expressed as a combination of flags, including zero.

BadEnum can still show up from time to time, and in the case of flags it will be
for a sum that can not be represented with known flags

Reviewed By: ambv

Differential Revision: D6564348

fbshipit-source-id: 408e98db8e21266f67615660aab61c8e1ac7a9ee